MrImage.plot - bugfix such that created files for spmi are properly deleted
git-svn-id: http://chap04.ethz.ch/fmri_svn/development@424 af1c8887-3614-4b1b-9890-8700dbf8cd7f

+++ analysis/code/classes/@MrImage/plot.m
@@ -445,11 +445,13 @@ else % different plot types: montage, 3D, spm
             fileName = this.parameters.save.fileName;
             fileNameNifti = fullfile(this.parameters.save.path, ...
                 regexprep(fileName, '\..*$', '\.nii'));
-            
-            % create nifti file, if not existing
+            doDelete = false;
+            % create nifti file, if not existing and take note to delete it
+            % afterwards
             % TODO: how about saved objects with other file names
             if ~exist(fileNameNifti, 'file')
                 this.save('fileName', fileNameNifti);
+                doDelete = true;
             end
             
             % select Volumes
@@ -470,8 +472,13 @@ else % different plot types: montage, 3D, spm
                         input('Press Enter to leave interactive mode');
                 end
                 
-                delete(fileNameNifti);
-                [stat, mess, id] = rmdir(this.parameters.save.path);
+                if doDelete
+                    delete(fileNameNifti);
+                    [fp, fn] = fileparts(fileNameNifti);
+                    fileNameDimInfo = fullfile(fp, [fn '_dimInfo.mat']);
+                    delete(fileNameDimInfo);
+                    [stat, mess, id] = rmdir(this.parameters.save.path);
+                end
             end
             
             % restore original file name
